VRS: Refactor performance metrics as common methods
